requirements.txt

requests==2.28.2
beautifulsoup4==4.12.2
scraperapi-sdk==1.0.0
python-dotenv==1.0.0 # For handling .env file
